moved LuminHandDataConverter to Lumin.Utilities
