+++ XRTK.Lumin/Packages/com.xrtk.lumin/Controllers/LuminControllerDataProvider.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
 using XRTK.Lumin.Profiles;
+using XRTK.Lumin.Utilities;
 using XRTK.Providers.Controllers;
 
 #if PLATFORM_LUMIN
