+++ XRTK.Lumin/Packages/com.xrtk.lumin/Utilities.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 14fb9301f77d904429c2851db462d235
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
